==> 01_EDA/01_Understanding.ipynb <==
# !pip install pandas

import pandas as pd

# 1. Do not disturb the Original Data 
# 2. DO not edit anything on the data manually

df = pd.read_csv('../00_Data/data.csv')
df.head()

pd.set_option('display.max_columns', None)

df.head()

# List out All Columns
df.columns

1. Salary Date

df['Salary Date'].describe()

# Date Range {m/d/y}
df['Salary Date'].min(), df['Salary Date'].max()

2. Salary Month

df['Salary Month'].describe()

3. Salary ID

df['Salary ID'].describe()

Salary Status

df['Salary Status'].describe()

df['Salary Status'].unique()

df['Salary Status'].value_counts()

Employee ID

df['Employee ID'].describe()

# Memory Usage

df.memory_usage(deep=True)

# Change the Category Type
df['Employee ID'] = df['Employee ID'].astype('category')
df.memory_usage(deep=True)

# Change to String Type
df['Employee ID'] = df['Employee ID'].astype('string')
df.memory_usage(deep=True)

Best Choice - Category Type

# Change the Category Type
df['Employee ID'] = df['Employee ID'].astype('category')

df['Employee ID'].describe()

# We can understand from unique values, there are row which are repeating 
# multiple times for the same employees due to every month salary
# There are totally 2285 employees

Employee Name

df['Employee Name'].describe()

Gender

df['Gender'].describe()

df['Gender'].value_counts(normalize=True) * 100

# Totally 79% Male Employees and  21% Female Employees are working

Generation

df['Generation'].describe()

df['Generation'].unique()

df['Generation'].value_counts()

Joining

df['Joining'].describe()

Termination

df['Termination'].describe()

# Employee terminated date given, 9 employees max resigned in a same day
# Joining and Termination are important columns for our prediction

Term Category

df['Term Category'].describe()

df['Term Category'].value_counts()

Term Reason

df['Term Reason'].describe()

Term Sub Reason

df['Term Sub Reason'].describe()

df['Term Sub Reason'].unique()

Employee Grade

df['Employee Grade'].describe()

df['Employee Grade'].value_counts()

Location

df['Location'].describe()

df['Location'].value_counts()

Career Bucket

df['Career Bucket'].describe()

df['Career Bucket'].value_counts()

Age Bucket

df['Age Bucket'].describe()

df['Age Bucket'].value_counts()

Manage ID

df['Manager ID'].describe()

df['Manager ID'] = df['Manager ID'].astype('category')
df['Manager ID'].describe()

Manager Name

df['Manager Name'].describe()

Performance

df['Performance'].describe()

df['Performance'].value_counts()

TALENT

df['Talent'].describe()

df['Talent'].value_counts()

Department

df['Department'].describe()

df['Department'].value_counts()

df.head()
