==> 01_EDA/02_exploratory_data_analysis.ipynb <==
Exploratory Data Analysis

# !pip install --upgrade pip
# !pip install ydata-profiling
# !pip install --upgrade setuptools

from ydata_profiling import ProfileReport
import pandas as pd
import time

df = pd.read_csv('../00_Data/data.csv')
df.head()

profile = ProfileReport(df, title='Test title', html = {'style':{'full_width': True}})
profile.to_file(output_file='profile_report')

round(time.time() * 1000)

def generate_profile_report(df, path, title):
    profile = ProfileReport(df, title=title, html = {'style':{'full_width': True}})

    epoch = round(time.time() * 1000)
    report_name = path + '/profile_' + str(epoch) + '.html'
    profile.to_file(output_file=report_name)


generate_profile_report(df, 'reports', 'Original Data')
