==> 01_EDA/03_ValidateTypes.ipynb <==
Type Validation

1. Constant Features => Column has only one value
2. Binary Features => Column has two values
3. Categorical => More than two unique categorical (text) values
4. Numerical => More than two unique number values

import pandas as pd 

df = pd.read_csv('../00_Data/data.csv')
df.head()

df.columns

df['Salary Date'].nunique()

def validate_types(df:pd.DataFrame):
    """Tabulate Data Types and Unique Values"""
    columns = df.columns
    col_types = []

    for column in columns:
        data_type = 'object'

        if df[column].nunique() > 2:
            if df[column].dtype == 'int64' or df[column].dtype == 'float64':
                data_type = 'numerical'
            else:
                data_type = 'categorical'
        elif df[column].nunique() > 1:
            data_type = 'binary'
        else:
            data_type = 'constant'

        # Total Unique Values
        nuniques = df[column].nunique()

        col_types.append([column, data_type, nuniques])

    df = pd.DataFrame(col_types)
    df.columns = ['Column', 'Data Type', 'Unique Values']
    df = df.sort_values(by='Unique Values', ascending=False)
    return df


validate_types(df)
