==> 01_EDA/04_MissingValues.ipynb <==
Missing values

import pandas as pd

df = pd.read_csv('../00_Data/data.csv')
df.head()

for series in df.isnull().sum().items():
    print(series[0], series[1])

df.shape[0]

def missing_percent(df:pd.DataFrame):
    """Tabulate Missing Percentage"""

    nulls = df.isnull().sum()
    total_rows = df.shape[0]

    missing_list = []
    for series in nulls.items():
        if series[1]> 0:
            percent = (series[1]/ total_rows) * 100
            missing_list.append([series[0], series[1], percent])

    df_missing = pd.DataFrame(missing_list)
    df_missing.columns = ['Column', 'Missing Values', 'Percentage']
    df_missing = df_missing.sort_values(by='Missing Values', ascending=False)
    return df_missing

missing_percent(df)
