==> 01_EDA/05_LatestRecords.ipynb <==
Optimizing Rows | Considering Latest Records only

import pandas as pd

df = pd.read_csv('../00_Data/data.csv')
df.head()

df.shape

df.loc[df['Employee ID'] == 100001]

df.dtypes

df['Salary Date'] = pd.to_datetime(df['Salary Date'])
df.dtypes

df_latest = df.loc[df['Salary Date'] == df.groupby('Employee ID')['Salary Date'].transform('max')]
df_latest

df_latest.shape

df_latest.to_csv('../00_Data/01_LatestRecords.csv', index=False)
