==> 01_EDA/06_ResignedEmployees.ipynb <==
Inactive or Resigned Employees

import pandas as pd
from datetime import datetime

df = pd.read_csv('../00_Data/01_LatestRecords.csv')
df.head()

df['Joining'].dtype, df['Termination'].dtype

df['Joining'] = pd.to_datetime(df['Joining'])
df['Termination'] = pd.to_datetime(df['Termination'])

df.dtypes

def calculate_active_months(start_date:datetime, end_date:datetime):
    """CALCUALTE ACTIVE MONTHS"""
    months = (end_date.year - start_date.year)*12 + (end_date.month - start_date.month)
    return months


df_inactive = df.copy()
df_inactive.loc[:, 'active_months'] = df_inactive.apply(lambda x: 
                                                        calculate_active_months(x['Joining'], 
                                                                                x['Termination']), axis=1)
df_inactive.head()                                                                         

df_inactive = df_inactive.loc[df['Salary Status'] == 'Inactive']

df_inactive.to_csv('../00_Data/02_ResignedEmployees.csv', index=False)

df_inactive.shape
