==> 01_EDA/07_OptimizeColumns.ipynb <==
Optimize Columns

import sys, os
project_libs = os.path.abspath('../libs')
sys.path.append(project_libs)

import pandas as pd
from InputOptimization import InputOptimization

io = InputOptimization()

df = pd.read_csv('../00_Data/02_ResignedEmployees.csv')
df.head()

df.shape

io.generate_profile_report(df, '../01_EDA/reports', 'Resigned Employees')

Remove Columns

1. Salary Month
2. Salary ID
3. Salary Status
4. Em ployee ID
5. Employee Name
6. Term Category
7. Term Reason
8. Manager Name
9. Performance


def remove_columns(df:pd.DataFrame, columns:list):
    df = df.drop(columns=columns, axis=1)
    return df

df.columns

rem_cols = ['Salary Month', 'Salary ID', 'Salary Status',
            'Employee ID', 'Employee Name', 'Term Category',
            'Term Reason', 'Manager Name', 'Performance', 'Salary Date', 'Joining', 'Termination']
df = remove_columns(df, rem_cols)
df.head()

df.to_csv('../00_Data/03_OptimizedColumns.csv', index=False)

io.generate_profile_report(df, '../01_EDA/reports/', 'Optimized Columns')
