==> 01_EDA/08_Outliers.ipynb <==
Outliers Handling

import pandas as pd

df = pd.read_csv('../00_Data/03_OptimizedColumns.csv')
df.head()

df.shape

Visual Understanding

df['active_months'].sort_values().unique()

df = df.loc[df['active_months'] < 200]
df.shape

Inter-Quartile Range (IQR)

info = df['active_months'].describe()
info

q3 = info.loc['75%']
q1 = info.loc['25%']

iqr = q3-q1
iqr

upper_bound = q3 + (1.5 * iqr)
lower_bound = q1 - (1.5 * iqr)

upper_bound, lower_bound

Outlier Percentage

# Outlier Rows

df_upper_rows = df[df['active_months'] >= upper_bound]
df_lower_rows = df[df['active_months'] <= lower_bound]

outlier_rows = df_upper_rows.shape[0] + df_lower_rows.shape[0]
total_rows = df.shape[0]

outlier_percent = round((outlier_rows/total_rows)* 100, 1)
outlier_percent

Non Outlier Rows

df_nout = df.copy()
df_nout = df_nout[df_nout['active_months'] < upper_bound]
df_nout = df_nout[df_nout['active_months'] > lower_bound]
df_nout.shape

df_nout.to_csv('../00_Data/04_NoOutliers.csv', index=False)

Outlier Rows

df_upper = df[df['active_months'] >= upper_bound]
df_lower = df[df['active_months'] <= lower_bound]

df_out = pd.concat([df_upper, df_lower])
df_out.shape

df_out

df_out.to_csv('../00_Data/Outliers.csv', index=False)
