==> 01_EDA/09_Encoding.ipynb <==
Data Encoding - Conversion of Categorical Data Into Numerical Data

# !pip install scikit-learn

import sys, os
lib_path = os.path.abspath('../libs')
sys.path.append(lib_path)

import pandas as pd
from sklearn.preprocessing import LabelEncoder

from InputOptimization import InputOptimization

io = InputOptimization()

df = pd.read_csv('../00_Data/04_NoOutliers.csv')
df.head()

df.shape

io.validate_types(df)

Encoding Binary Columns

df['Gender'].unique()

def binary_encoding(df:pd.DataFrame, column:str, true_value:str):
    """Binary Encoding"""

    unique_values = df[column].unique()

    for val in unique_values:
        if val == true_value:
            df.loc[:, column] = df[column].replace(val, 1).infer_objects(copy=False)
        else:
            df.loc[:, column] = df[column].replace(val, 0).infer_objects(copy=False)

    return df

df = binary_encoding(df, 'Gender', 'Female')
df.head()

df['Talent'].unique()

Columns with more than 2 Unique Categorical Values

1. Label Encoding

label_cols = ['Term Sub Reason', 'Location', 'Talent', 'Department']

# Label Encoding - If values are not in order
def label_encoding(df:pd.DataFrame, columns:list):
    """Label Encoding Function"""

    le = LabelEncoder()
    for column in columns:
        df[column] = le.fit_transform(df[column])
        df[column] = df[column] + 1
        df[column] = pd.to_numeric(df[column])
    
    return df

df['Term Sub Reason'].unique()

df = label_encoding(df, label_cols)
df.head()

df['Term Sub Reason'].unique()

2. Ordinal Encoding - For Ordered Values

df.head()

df['Generation'].unique()

def ordinal_encoding(df:pd.DataFrame, column:str, order_list:list):
    """Encoding Ordered Data"""

    data = df.copy()
    count = 1
    for item in order_list:
        data[column] = data[column].replace(item, count)
        count = count +1 

    data[column] = pd.to_numeric(data[column])
    return data

gen_list = ['Gen Z', 'Millennials', 'Gen X', 'Baby Boomers']
df = ordinal_encoding(df, 'Generation', gen_list)
df.head()

df['Employee Grade'].unique()

gen_list = ['Junior', 'Mid', 'Senior']
df = ordinal_encoding(df, 'Employee Grade', gen_list)
df.head()

df['Career Bucket'].unique()

gen_list = ['Below 30', '31 to 35', 'Above 35']
df = ordinal_encoding(df, 'Career Bucket', gen_list)
df.head()

df['Age Bucket'].unique()

order_list = ['21 To 25', '26 To 30', '31 To 35', '36 T0 40',  '41 T0 45', '46 T0 50', '51 T0 55']
df = ordinal_encoding(df, 'Age Bucket', order_list)
df.head()

df.to_csv('../00_Data/05_Encoded.csv', index=False)
