==> 01_EDA/10_Normalization.ipynb <==
Normalization

import sys, os
lib_path = os.path.abspath('../libs')
sys.path.append(lib_path)

from InputOptimization import InputOptimization

io = InputOptimization()

import pandas as pd
from sklearn.preprocessing import StandardScaler, MinMaxScaler, RobustScaler

df = pd.read_csv('../00_Data/05_Encoded.csv')
df.head()

x + 2y + z + p + q + r + 2a + 102292b + 5c + 2d = 59.0

Scaling:

1. Standard Scaler | Z Scaler
2. Min-Max Scaler 
3. Robust Scaler

1. Standard Scaler

df['Manager ID'].values.reshape(-1, 1)

def std_scale(df:pd.DataFrame, columns:list):
    """Scale the numerical column using Standard Scaler"""

    df_std = df.copy()
    scaler = StandardScaler()

    for column in columns:
        df_std[column] = scaler.fit_transform(df_std[column].values.reshape(-1, 1))
        df_std[column] = df_std[column].apply(lambda x: round(x, 1))

    return df_std

df_std = std_scale(df, ['Manager ID'])
df_std.head()

2. Min-Max Normalization

def min_max_scale(df:pd.DataFrame, columns:list):
    """Scaling numeric columns using Min-Max Normalization"""

    df_minmax = df.copy()
    scaler = MinMaxScaler()

    for column in columns:
        df_minmax[column] = scaler.fit_transform(df_minmax[column].values.reshape(-1, 1))
        df_minmax[column] = df_minmax[column].apply(lambda x: round(x, 1))

    return df_minmax


df_minmax = min_max_scale(df, ['Manager ID'])
df_minmax.head()

3. Robust Scaler

def robust_scaler(df:pd.DataFrame, columns:list):
    """Scaling numeric columns using Robust Normalization"""

    df_robust = df.copy()
    scaler = RobustScaler()

    for column in columns:
        df_robust[column] = scaler.fit_transform(df_robust[column].values.reshape(-1, 1))
        df_robust[column] = df_robust[column].apply(lambda x: round(x, 1))

    return df_robust

df_robust = robust_scaler(df, ['Manager ID'])
df_robust.head()

General Validation

# Profile Report
io.generate_profile_report(df_std, '../01_EDA/reports/', 'Normalized')

# Validate Type
io.validate_types(df_std)

df_std
