==> 01_EDA/12_ModelOptimization.ipynb <==
import sys, os
lib_path = os.path.abspath('../libs')
sys.path.append(lib_path)

Model Optimization

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_absolute_error, mean_squared_error

import math

from InputOptimization import InputOptimization

df = pd.read_csv('../00_Data/05_Encoded.csv')
df.head()

df.drop_duplicates(inplace=True)

io = InputOptimization()

io.generate_profile_report(df, '../01_EDA/reports/', 'Duplicates Removed')

y = m1x1 + m2x2 + ....

df.head()

df.corr()

fig = plt.figure(figsize=(9, 6))
ax = fig.add_subplot()
plt.show()

df['Generation'].corr(df['Employee Grade']) *100

df.columns[:]

def correlation_table(df:pd.DataFrame):
    correlations = []
    count = 1
    for column1 in df.columns:
        for column2 in df.columns[count:]:
            corr = df[column1].corr(df[column2])
            correlations.append([column1, column2, corr])

        count = count + 1

    df_correlation = pd.DataFrame(correlations)
    df_correlation.columns = ['Feature 1', 'Feature 2', 'Correlation']
    df_correlation = df_correlation.sort_values(by='Correlation', ascending=False)
    return df_correlation

correlation_table(df).head()

Regression Function

list(range(1, 11))

def linear_regression(df:pd.DataFrame, iteration=1):

    result = []
    for num in range(1, iteration+1):

        X = df.iloc[:, :-1]
        y = df.iloc[:, -1]

        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

        scaler = StandardScaler()
        XS_train = scaler.fit_transform(X_train)
        XS_test = scaler.transform(X_test)

        model = LinearRegression()
        model.fit(XS_train, y_train)

        # Testing Model
        y_pred = model.predict(XS_test)
        mse = mean_squared_error(y_test, y_pred)
        rmse = math.sqrt(mse)
        mae = mean_absolute_error(y_test, y_pred)
        r2 = model.score(XS_test, y_test) * 100

        result.append([model, mse, rmse, mae, r2])
    
    df_result = pd.DataFrame(result)
    df_result.columns = ['Model', 'MSE', 'RMSE', 'MAE', 'R2']
    df_result = df_result.sort_values('R2', ascending=False)

    return df_result
        

df.drop(['Career Bucket', 'Generation', 'Employee Grade'], axis=1, inplace=True)

correlation_table(df).head(5)

lm_models = linear_regression(df, 100)
lm_models

lm_models.loc[94]['Model']

df.head()

df.to_csv('../00_Data/06_CorrelationDrop.csv', index=False)
