==> 01_EDA/14_Model Splitting.ipynb <==
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import classification_report

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df.head()

def classification(df:pd.DataFrame, iteration=1):

    result = []
    for num in range(1, iteration+1):
        X = df.iloc[:, :-1]
        y = df.iloc[:, -1]

        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

        scaler = StandardScaler()
        XS_train = scaler.fit_transform(X_train)
        XS_test = scaler.transform(X_test)

        model = KNeighborsClassifier()
        model.fit(XS_train, y_train)

        train_score = round(model.score(XS_train, y_train) * 100, 2)
        test_score = round(model.score(XS_test, y_test) * 100, 2)

        stability = abs(train_score - test_score) 

        result.append([model, train_score, test_score, stability])
    
    df_result = pd.DataFrame(result)
    df_result.columns = ['Model', 'Train Score', 'Test Score', 'Stability']
    df_result = df_result.sort_values(by='Stability')

    return df_result

classification(df, 100).head(10)

def classify_active_months(num):

    if num < 48:
        value = 'ST'
    elif num < 96:
        value = 'MT'
    else:
        value = 'LT'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100).head(10)

1. Model - LTO

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df.head()

def classify_active_months(num):

    if num > 144:
        value = 'LT'
    else:
        value = 'OTH'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100).head(10)

2. Model - LT2 (ABOVE 10 YEARS)

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df = df[df['active_months'] < 144]
df.head()

def classify_active_months(num):

    if num > 120:
        value = 'LT2'
    else:
        value = 'OTH'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100)

3. Model - LT1 (ABOVE 8 YEARS)

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df = df[df['active_months'] < 120]
df.head()

def classify_active_months(num):

    if num > 96:
        value = 'LT1'
    else:
        value = 'OTH'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100).head(10)

4. Model - MT2 (ABOVE 6 YEARS)

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df = df[df['active_months'] < 96]
df.head()

def classify_active_months(num):

    if num > 72:
        value = 'MT2'
    else:
        value = 'OTH'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100).head(10)

5. Model - MT1 (ABOVE 4 YEARS)

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df = df[df['active_months'] < 72]
df.head()

def classify_active_months(num):

    if num > 48:
        value = 'MT1'
    else:
        value = 'OTH'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100).head(10)

6. Model - ST2 (ABOVE 2 YEARS)

df = pd.read_csv('../00_Data/06_CorrelationDrop.csv')
df = df[df['active_months'] < 48]
df.head()

def classify_active_months(num):

    if num > 12:
        value = 'ST2'
    else:
        value = 'ST1'

    return value

df['active_months'] = df['active_months'].apply(lambda x:classify_active_months(x))
df.head(3)

classification(df, 100).head(10)
